f/mmcdaniel-saturation: Redefine HbO2 and HbCO2 species
Previous: HbO2 and HbCO2 were hemoglobin with O2/CO2 bound.  This
definition required us to track HbO2CO2 as well (both gases bound to one
Hb) in order to get enough gas in the system.  However, the only
information the saturation solver needs is amount of bound O2/CO2,
regardless of how it is bound (by itself or with the other gas).
Furthermore, total Hb is based on the estimate for HbO2CO2, and this
introduces the possibility of not conserving total hemoblobin at extreme
PO2/PCO2 values.

Solution: Re-define HbO2 and HbCO2 to mean "bound O2" and "bound CO2"
and Hb to mean "total Hb". Total Hb should be constant over a run,
except in pathophysiological states (e.g. fluid shift in sepsis).
Calculations are not affected--this removes the need to estimate HbO2CO2
and removes possibility of not conserving Hb.